# src/data/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AWSSBG Internal Dashboard API",
    "description": "REST API for the AWS Student Builder Group Internal Dashboard — members, tasks, submissions, link shortener, audit logs, and analytics. All endpoints (except auth) require a valid `sbg_session` HttpOnly cookie obtained via the OTP login flow.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Relative to deployment origin" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sbg_session"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-secret",
        "description": "Shared secret configured via the CRON_SECRET env var. An `Authorization: Bearer <secret>` header is also accepted."
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["SBG_LEADER", "SECRETARY", "DIRECTOR", "MANAGER", "ASSOCIATE", "BUILDER"]
      },
      "Domain": {
        "type": "string",
        "enum": ["Technical", "Corporate", "Creatives", "General"]
      },
      "Subdomain": {
        "type": "string",
        "enum": ["Software Development", "AI & Machine Learning", "Cloud & DevOps", "Events & Operations", "Sponsorship & Finance", "HR & Admin", "PR & Marketing", "Digital Design", "Media Production"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "SessionUser": {
        "type": "object",
        "properties": {
          "memberId": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/Role" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true }
        }
      },
      "Member": {
        "type": "object",
        "properties": {
          "memberId": { "type": "string" },
          "clubId": { "type": "string" },
          "name": { "type": "string" },
          "regNo": { "type": "string" },
          "department": { "type": "string" },
          "section": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "officialEmail": { "type": "string", "format": "email" },
          "personalEmail": { "type": "string" },
          "phone": { "type": "string" },
          "whatsapp": { "type": "string" },
          "github": { "type": "string" },
          "linkedin": { "type": "string" },
          "instagram": { "type": "string" },
          "meetup": { "type": "string" },
          "builderId": { "type": "string" },
          "faName": { "type": "string" },
          "faEmail": { "type": "string" },
          "faPhone": { "type": "string" },
          "joinedAt": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean" },
          "totalStars": { "type": "number" }
        }
      },
...
      "MemberUpdateRequest": {
        "type": "object",
        "description": "Self-service callers may only set phone, github, linkedin, meetup, personalEmail. Presidium may additionally set name, role, domain, subdomain, department, builderId, isActive, clubId, regNo.",
        "properties": {
          "name": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "domain": { "$ref": "#/components/schemas/Domain" },
          "subdomain": { "$ref": "#/components/schemas/Subdomain" },
          "department": { "type": "string" },
          "phone": { "type": "string" },
          "github": { "type": "string" },
          "linkedin": { "type": "string" },
          "meetup": { "type": "string" },
          "builderId": { "type": "string" },
          "personalEmail": { "type": "string" },
          "isActive": { "type": "boolean" },
          "clubId": { "type": "string" },
          "regNo": { "type": "string" }
        }
      },
...
      "Rating": {
        "type": "object",
        "properties": {
          "memberId": { "type": "string" },
          "memberName": { "type": "string" },
          "domain": { "$ref": "#/components/schemas/Domain", "nullable": true },
          "subdomain": { "$ref": "#/components/schemas/Subdomain", "nullable": true },
          "role": { "$ref": "#/components/schemas/Role" },
          "totalStars": { "type": "number" },
          "approvedCount": { "type": "integer" },
          "lateApprovedCount": { "type": "integer" },
          "rejectedCount": { "type": "integer" },
          "pendingCount": { "type": "integer" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "No valid session cookie present",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Unauthorized" } } }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted to perform this action",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Forbidden" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "sessionCookie": [] }],
  "paths": {
    "/auth/send-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Send a login OTP to a member's official email",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["email"], "properties": { "email": { "type": "string", "format": "email" } } } } }
        },
        "responses": {
          "200": { "description": "OTP sent", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } } } } } },
          "400": { "description": "Missing email" },
          "403": { "description": "Account is inactive" },
          "404": { "description": "Email not registered" },
          "429": { "description": "Resend cooldown in effect" }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify OTP and create a session",
        "description": "On success, sets the `sbg_session` HttpOnly cookie.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["email", "otp"], "properties": { "email": { "type": "string", "format": "email" }, "otp": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Session created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "user": { "$ref": "#/components/schemas/SessionUser" } } } } } },
          "400": { "description": "Missing email or OTP" },
          "401": { "description": "Invalid or expired OTP" },
          "403": { "description": "Account is inactive" },
          "404": { "description": "Member not found" },
          "429": { "description": "Too many failed attempts" }
        }
      }
    },
...
    "/members/{memberId}": {
      "get": {
        "tags": ["Members"],
        "summary": "Get a member by id (includes rating)",
        "parameters": [{ "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Member detail", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "allOf": [{ "$ref": "#/components/schemas/Member" }, { "type": "object", "properties": { "rating": { "$ref": "#/components/schemas/Rating", "nullable": true } } }] } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Members"],
        "summary": "Update a member",
        "description": "Presidium may edit any field on any member. Non-presidium callers may only edit their own record, and only: phone, github, linkedin, meetup, personalEmail.",
        "parameters": [{ "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemberUpdateRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "No valid fields to update" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["Members"],
        "summary": "Deactivate a member (soft delete)",
        "description": "Presidium only. Sets isActive=false; does not remove the record.",
        "parameters": [{ "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deactivated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  }
}